requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
cachetools==5.3.0

charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
cycler==0.11.0
datasets==1.18.4
decorator==5.1.1
dill==0.3.6
executing==1.2.0
filelock==3.10.6
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
geomloss==0.2.6
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
huggingface-hub==0.13.3
idna==3.4
img2vec-pytorch==1.0.1
importlib-metadata==6.6.0
ipdb==0.13.13
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
jiwer==3.0.1
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.5
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.22.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.0
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
protobuf==4.22.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
rapidfuzz==2.13.7
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
sacremoses==0.0.53
scikit-learn==1.1.0
scipy==1.10.1
seqeval==1.2.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1

tqdm==4.64.0
traitlets==5.9.0
transformers==4.17.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.3
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
